Initial nix expression generator.
This is *not* ready for external use, primarily because of the
hardcoded `justStaticExecutables` case and the call to my local
cabal2nix program. Some of the code may also break due to upstream
cabal2nix or stack changes, and not all the features of either tool
are reflected here. All aspects of the API are still in flux.

With those caveats in mind, I can successfully generate and build nix
expressions for my test project:

    $ git clone https://github.com/jmitchell/haskell-multi-package-demo1
    $ cd haskell-multi-package-demo1
    $ stack2nix stack.yaml
    $ nix-build -A haskell-multi-proj-demo1-static