Generate --subpath for local targets
Some packages may need access to files outside of their subdirectory
(e.g. hpack files may source other hpack files). Therefore it is
important to keep the integrity of the project directory and specify
individual subdirectories by passing `--subpath` to `cabal2nix`.

Since `stack` does not support specifying subdirs for file paths (unlike
git repositories), it is impossible to tell in general where is the root
of the dependency. But we can at least do this for local targets.